requirements.txt
dj-database-url==0.5.0
Django==2.0.6
django-appconf==1.0.2
django-carton==1.2.1
django-imagekit==4.0.2
django-widget-tweaks==1.4.2
pilkit==2.0
Pillow==5.2.0
psycopg2==2.7.5
python-decouple==3.1
six==1.11.0
sqlparse==0.2.4